requirements.txt
-e .
reppy==0.4.14
feedparser==5.2.1
requests==2.23.0
lxml==4.5.1
newspaper3k==0.2.8
pymongo==3.10.1